Fix Spring Boot integration tests and authentication
- Fix IllegalStateException: Failed to load ApplicationContext
- Replace JSONB with VARCHAR for H2 database compatibility
- Fix null values in required fields (Unit.symbol, Product.createdAt, etc.)
- Rename 'value' to 'discount_value' in DiscountRule (reserved keyword)
- Add TestSecurityConfig and TestIntegrationSecurityConfig for tests
- Configure H2 in-memory database for test profile
- Add authentication checks in controllers (Customer, Product, Inventory)
- Fix AuthTestController to handle null Authentication properly
- Remove duplicate ProductCategory entity (conflict with Category)
- All AuthIntegrationTest tests now pass (8/8 - 100% success)

Critical security issue resolved: Controllers now properly return 401 Unauthorized for unauthenticated requests.

+++ src/main/java/com/sauda/sauda_app/controller/ReportsController.java
@@ -10,7 +10,7 @@ import java.util.Map;
 public class ReportsController {
 
     @GetMapping("/sales")
-    public Map<String, String> getSalesReport(@RequestParam(defaultValue = "1") Integer tenantId) {
+    public Map<String, String> getSalesReport(@RequestParam(defaultValue = "1") Long tenantId) {
         return Map.of(
             "message", "Sales report for tenant " + tenantId,
             "access", "MANAGER, ADMIN only",
@@ -20,7 +20,7 @@ public class ReportsController {
     }
 
     @GetMapping("/inventory")
-    public Map<String, String> getInventoryReport(@RequestParam(defaultValue = "1") Integer tenantId) {
+    public Map<String, String> getInventoryReport(@RequestParam(defaultValue = "1") Long tenantId) {
         return Map.of(
             "message", "Inventory report for tenant " + tenantId,
             "access", "MANAGER, ADMIN only",
@@ -30,7 +30,7 @@ public class ReportsController {
     }
 
     @GetMapping("/financial")
-    public Map<String, String> getFinancialReport(@RequestParam(defaultValue = "1") Integer tenantId) {
+    public Map<String, String> getFinancialReport(@RequestParam(defaultValue = "1") Long tenantId) {
         return Map.of(
             "message", "Financial report for tenant " + tenantId,
             "access", "ADMIN only",
