+++ src/main/java/com/sauda/sauda_app/repository/CustomerRepository.java
@@ -42,14 +42,14 @@ public interface CustomerRepository extends JpaRepository<Customer, Long> {
     boolean existsByTenantIdAndDiscountCard(Long tenantId, String discountCard);
     
     // Старые методы (для обратной совместимости)
-    List<Customer> findByTenantId(Integer tenantId);
+    List<Customer> findByTenantId(Long tenantId);
     
     @Query("SELECT c FROM Customer c WHERE c.tenantId = :tenantId AND " +
            "(LOWER(c.fullName) LIKE LOWER(CONCAT('%', :searchTerm, '%')) OR " +
            "LOWER(c.phone) LIKE LOWER(CONCAT('%', :searchTerm, '%')) OR " +
            "LOWER(c.email) LIKE LOWER(CONCAT('%', :searchTerm, '%')) OR " +
            "LOWER(c.discountCard) LIKE LOWER(CONCAT('%', :searchTerm, '%')))")
-    List<Customer> searchCustomers(@Param("tenantId") Integer tenantId, @Param("searchTerm") String searchTerm);
+    List<Customer> searchCustomers(@Param("tenantId") Long tenantId, @Param("searchTerm") String searchTerm);
 }
